==> 03_base_acertos.ipynb <==
import pandas as pd 

# Comparação simulado EXATAS

A última etapa que foi planjeada foi realizar a comparaçãp dos simulados de exatas, já que não teve outro de humanas para que fosse realizado a comparação.

### Carregamento das bases

simulado1 = pd.read_csv('data/serving/simulado_final_1.2.csv')
simulado3 = pd.read_csv('data/serving/simulado_final_2.2.csv')

### Seleção das colunas alvo

simulado3 = simulado3[['ESTUDANTE','QTD_ACERTOS','ACERTOS_CN','ACERTOS_MT']]
simulado1 = simulado1[['ESTUDANTE','QTD_ACERTOS','ACERTOS_CN','ACERTOS_MT']]

### Seleção dos estudantes que participaram de ambos os simulados

estudantes_simulado = list(simulado1.ESTUDANTE.unique())+list(simulado3.ESTUDANTE.unique())
estudantes_simulado = set(estudantes_simulado)

df_estudantes = pd.DataFrame(estudantes_simulado)
df_estudantes.columns = ['ESTUDANTE']

### Unindo os datasets com base na lista de estudantes únicos

df_estudantes = df_estudantes.merge(simulado1,how='left')
df_estudantes.columns = ['ESTUDANTE', 'QTD_ACERTOS_EXATAS_1', 'ACERTOS_CN_EXATAS_1', 'ACERTOS_MT_EXATAS_1']

df_estudantes = df_estudantes.merge(simulado3,how='left',left_on='ESTUDANTE',right_on='ESTUDANTE')

df_estudantes.columns = ['ESTUDANTE', 'QTD_ACERTOS_EXATAS_1', 'ACERTOS_CN_EXATAS_1',
       'ACERTOS_MT_EXATAS_1', 'QTD_ACERTOS_EXATAS_2', 'ACERTOS_CN_EXATAS_2', 'ACERTOS_MT_EXATAS_2']

### Removendo os alunos que não realizaram ambos os simulados

compativo_simulado_exatas = df_estudantes.dropna()


### Definindo como inteiro as colunas de acertos

colunas_acertos = ['QTD_ACERTOS_EXATAS_1', 'ACERTOS_CN_EXATAS_1', 'ACERTOS_MT_EXATAS_1',
                          'QTD_ACERTOS_EXATAS_2', 'ACERTOS_CN_EXATAS_2', 'ACERTOS_MT_EXATAS_2']

compativo_simulado_exatas[colunas_acertos] = compativo_simulado_exatas[colunas_acertos].astype(int)

### Exportando a base de dados

# compativo_simulado_exatas.to_csv('comparativo_simulado_exatas.csv', index=False)
